requirements.txt
annotated-types==0.7.0
blessed==1.20.0
botocore==1.37.17
chalice==1.31.4
click==8.1.8
dnspython==2.7.0
editor==1.6.6
iniconfig==2.1.0
inquirer==3.4.0
jmespath==1.0.1
packaging==24.2
pluggy==1.5.0
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pymongo==4.11.3
pytest==8.3.5
python-dateutil==2.9.0.post0
PyYAML==6.0.2
readchar==4.2.1
runs==1.2.2
setuptools==77.0.3
six==1.17.0
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
wheel==0.45.1
xmod==1.8.1
